:shirt: refactor: jest typescript
